requirements.txt
Django==1.7
Scrapy==0.24.4
Twisted==14.0.2
argparse==1.2.1
arrow==0.4.4
cffi==0.8.6
cryptography==0.6
cssselect==0.9.1
django-arrow-field==0.3.0
lxml==3.4.0
psycopg2==2.5.4
pyOpenSSL==0.14
pycparser==2.10
python-dateutil==2.2
queuelib==1.2.2
six==1.8.0
w3lib==1.10.0
wsgiref==0.1.2
zope.interface==4.1.1
